refactor os specific code to package platform
