--- surge/events_windows.go
@@ -1,12 +1,0 @@
-package surge
-
-func watchOSXHandler() {
-}
-
-func initOSHandler() {
-
-}
-
-func setVisualModeLikeOS() {
-
-}
