+++ surge/platform/events_windows.go
@@ -1,0 +1,15 @@
+package platform
+
+//WatchOSXHandler .
+func WatchOSXHandler() {
+}
+
+// InitOSHandler .
+func InitOSHandler() {
+
+}
+
+//SetVisualModeLikeOS .
+func SetVisualModeLikeOS() {
+
+}
